# tests/fixtures/studio_contracts/openapi/openapi_snapshot.json
{
  "components": {
    "schemas": {
...
      "CreateEpisodeRequest": {
        "additionalProperties": true,
        "properties": {
          "episode_number": {
            "default": 1,
            "minimum": 1.0,
            "title": "Episode Number",
            "type": "integer"
          },
          "idempotency_key": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Idempotency Key"
          },
          "metadata": {
            "additionalProperties": true,
            "title": "Metadata",
            "type": "object"
          },
          "schema_version": {
            "default": "studio.command/v1",
            "title": "Schema Version",
            "type": "string"
          },
          "series_id": {
            "title": "Series Id",
            "type": "string"
          },
          "title": {
            "minLength": 1,
            "title": "Title",
            "type": "string"
          }
        },
        "required": [
          "series_id",
          "title"
        ],
        "title": "CreateEpisodeRequest",
        "type": "object"
      },
...
      "CreateSeriesRequest": {
        "additionalProperties": true,
        "properties": {
          "description": {
            "default": "",
            "title": "Description",
            "type": "string"
          },
          "idempotency_key": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Idempotency Key"
          },
          "metadata": {
            "additionalProperties": true,
            "title": "Metadata",
            "type": "object"
          },
          "schema_version": {
            "default": "studio.command/v1",
            "title": "Schema Version",
            "type": "string"
          },
          "title": {
            "minLength": 1,
            "title": "Title",
            "type": "string"
          }
        },
        "required": [
          "title"
        ],
        "title": "CreateSeriesRequest",
        "type": "object"
      },
...
      "EpisodeListResponse": {
        "properties": {
          "items": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "title": "Items",
            "type": "array"
          }
        },
        "title": "EpisodeListResponse",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "RunEventsResponse": {
        "properties": {
          "after_sequence": {
            "title": "After Sequence",
            "type": "integer"
          },
          "events": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "title": "Events",
            "type": "array"
          },
          "latest_sequence": {
            "title": "Latest Sequence",
            "type": "integer"
          },
          "run_id": {
            "title": "Run Id",
            "type": "string"
          }
        },
        "required": [
          "run_id",
          "after_sequence",
          "latest_sequence"
        ],
        "title": "RunEventsResponse",
        "type": "object"
      },
      "RunResponse": {
        "properties": {
          "current_step": {
            "title": "Current Step",
            "type": "string"
          },
          "duration_sec": {
            "title": "Duration Sec",
            "type": "number"
          },
          "run_id": {
            "title": "Run Id",
            "type": "string"
          },
          "state": {
            "title": "State",
            "type": "string"
          },
          "step_count": {
            "title": "Step Count",
            "type": "integer"
          },
          "task_id": {
            "title": "Task Id",
            "type": "string"
          },
          "workflow_name": {
            "title": "Workflow Name",
            "type": "string"
          }
        },
        "required": [
          "run_id",
          "task_id",
          "workflow_name",
          "state",
          "current_step",
          "step_count",
          "duration_sec"
        ],
        "title": "RunResponse",
        "type": "object"
      },
...
      "SeriesListResponse": {
        "properties": {
          "items": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "title": "Items",
            "type": "array"
          },
          "next_cursor": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Next Cursor"
          }
        },
        "title": "SeriesListResponse",
        "type": "object"
      },
...
      "StartRunRequest": {
        "additionalProperties": false,
        "properties": {
          "schema_version": {
            "default": "studio.command/v1",
            "title": "Schema Version",
            "type": "string"
          }
        },
        "title": "StartRunRequest",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    }
  },
  "info": {
    "description": "Modular Monolith API V2 Production Application",
    "title": "WindAgent V2 API",
    "version": "0.3.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v2/runs/{run_id}": {
      "get": {
        "operationId": "get_run_api_v2_runs__run_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "run_id",
            "required": true,
            "schema": {
              "title": "Run Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RunResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Run",
        "tags": [
          "Runs V2"
        ]
      }
    },
...
    "/api/v3/studio/episodes/{episode_id}": {
      "get": {
        "operationId": "get_episode_api_v3_studio_episodes__episode_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "episode_id",
            "required": true,
            "schema": {
              "title": "Episode Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Get Episode Api V3 Studio Episodes  Episode Id  Get",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Episode",
        "tags": [
          "studio-v3",
          "studio-episodes"
        ]
      }
    },
...
    "/api/v3/studio/episodes/{episode_id}/runs": {
      "post": {
        "operationId": "start_or_resume_run_api_v3_studio_episodes__episode_id__runs_post",
        "parameters": [
          {
            "in": "path",
            "name": "episode_id",
            "required": true,
            "schema": {
              "title": "Episode Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "X-Idempotency-Key",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Idempotency-Key"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/StartRunRequest",
                "default": {
                  "schema_version": "studio.command/v1"
                }
              }
            }
          }
        },
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Start Or Resume Run",
        "tags": [
          "studio-v3",
          "studio-runs"
        ]
      }
    },
...
    "/api/v3/studio/runs/{run_id}": {
      "get": {
        "operationId": "get_run_api_v3_studio_runs__run_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "run_id",
            "required": true,
            "schema": {
              "title": "Run Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Get Run Api V3 Studio Runs  Run Id  Get",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Run",
        "tags": [
          "studio-v3",
          "studio-runs"
        ]
      }
    },
    "/api/v3/studio/runs/{run_id}/events": {
      "get": {
        "operationId": "get_run_events_api_v3_studio_runs__run_id__events_get",
        "parameters": [
          {
            "in": "path",
            "name": "run_id",
            "required": true,
            "schema": {
              "title": "Run Id",
              "type": "string"
            }
          },
          {
            "description": "Exclusive event sequence cursor",
            "in": "query",
            "name": "after",
            "required": false,
            "schema": {
              "default": 0,
              "description": "Exclusive event sequence cursor",
              "minimum": 0,
              "title": "After",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 100,
              "maximum": 1000,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RunEventsResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Run Events",
        "tags": [
          "studio-v3",
          "studio-runs"
        ]
      }
    },
    "/api/v3/studio/series": {
      "get": {
        "operationId": "list_series_api_v3_studio_series_get",
        "parameters": [
          {
            "in": "query",
            "name": "cursor",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Cursor"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 100,
              "maximum": 500,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SeriesListResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Series",
        "tags": [
          "studio-v3",
          "studio-series"
        ]
      },
      "post": {
        "operationId": "create_series_api_v3_studio_series_post",
        "parameters": [
          {
            "in": "header",
            "name": "X-Idempotency-Key",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Idempotency-Key"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateSeriesRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Series",
        "tags": [
          "studio-v3",
          "studio-series"
        ]
      }
    },
    "/api/v3/studio/series/{series_id}": {
      "get": {
        "operationId": "get_series_api_v3_studio_series__series_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "series_id",
            "required": true,
            "schema": {
              "title": "Series Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Get Series Api V3 Studio Series  Series Id  Get",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Series",
        "tags": [
          "studio-v3",
          "studio-series"
        ]
      }
    },
    "/api/v3/studio/series/{series_id}/episodes": {
      "get": {
        "operationId": "list_episodes_api_v3_studio_series__series_id__episodes_get",
        "parameters": [
          {
            "in": "path",
            "name": "series_id",
            "required": true,
            "schema": {
              "title": "Series Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EpisodeListResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Episodes",
        "tags": [
          "studio-v3",
          "studio-episodes"
        ]
      },
      "post": {
        "operationId": "create_episode_api_v3_studio_series__series_id__episodes_post",
        "parameters": [
          {
            "in": "path",
            "name": "series_id",
            "required": true,
            "schema": {
              "title": "Series Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "X-Idempotency-Key",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Idempotency-Key"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateEpisodeRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Episode",
        "tags": [
          "studio-v3",
          "studio-episodes"
        ]
      }
    },
...
  }
}